requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
chardet==4.0.0
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-heroku==0.0.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
filelock==3.3.1
idna==3.3
multidict==5.2.0
platformdirs==2.4.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
virtualenv==20.8.1
whitenoise==5.3.0
yarl==1.7.0
